+++ packages/alpha/src/types.ts
@@ -348,7 +348,7 @@ export interface SimulatorModelAggregateQuery {
 
 export interface SimulatorModelFilter {
   simulatorExternalIds?: CogniteExternalId[];
-  simulatorIntegrationExternalIds?: CogniteExternalId[];
+  externalIdPrefix?: string;
 }
 
 export interface SimulatorModelFilterQuery extends FilterQuery {