+++ packages/stable/src/__tests__/api/documentsApi.int.spec.ts
@@ -56,7 +56,7 @@ const getFileId = async (
   );
 };
 
-describe('Documents integration test', () => {
+describe('Documents integration test', { timeout: 3 * 60_000 }, () => {
   let client: CogniteClient;
   let fileId: number;
 
@@ -155,7 +155,7 @@ describe('Documents integration test', () => {
     expect(aggregate.groups[0].group[0].property).toStrictEqual(['labels']);
   });
 
-  describe('document preview', () => {
+  describe('document preview', { timeout: 3 * 60_000 }, () => {
     let documents: DocumentSearchResponse;
 
     beforeAll(async () => {
@@ -176,7 +176,8 @@ describe('Documents integration test', () => {
       await client.documents.preview.documentAsImage(document.id, 1);
     });
 
-    test(
+    // The "documentAsPdf" method currently takes more than 60s to respond, need to address this with backend team
+    test.skip(
       'fetch pdf preview',
       async () => {
         if (documents.items.length === 0) {
@@ -193,7 +194,7 @@ describe('Documents integration test', () => {
         const match = Buffer.from(frontSlice, 0).equals(Buffer.from(pdfPrefix));
         expect(match).toBe(true);
       },
-      30 * 1000
+      60 * 1000
     );
 
     test('fetch temporary link', async () => {