+++ packages/beta/CHANGELOG.md
@@ -3,6 +3,54 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [6.0.6](https://github.com/cognitedata/cognite-sdk-js/compare/@cognite/sdk-beta@6.0.5...@cognite/sdk-beta@6.0.6) (2025-09-29)
+
+**Note:** Version bump only for package @cognite/sdk-beta
+
+
+
+
+
+## [6.0.5](https://github.com/cognitedata/cognite-sdk-js/compare/@cognite/sdk-beta@6.0.4...@cognite/sdk-beta@6.0.5) (2025-09-25)
+
+**Note:** Version bump only for package @cognite/sdk-beta
+
+
+
+
+
+## [6.0.4](https://github.com/cognitedata/cognite-sdk-js/compare/@cognite/sdk-beta@6.0.3...@cognite/sdk-beta@6.0.4) (2025-09-22)
+
+**Note:** Version bump only for package @cognite/sdk-beta
+
+
+
+
+
+## [6.0.3](https://github.com/cognitedata/cognite-sdk-js/compare/@cognite/sdk-beta@6.0.2...@cognite/sdk-beta@6.0.3) (2025-09-19)
+
+**Note:** Version bump only for package @cognite/sdk-beta
+
+
+
+
+
+## [6.0.2](https://github.com/cognitedata/cognite-sdk-js/compare/@cognite/sdk-beta@6.0.0...@cognite/sdk-beta@6.0.2) (2025-09-17)
+
+**Note:** Version bump only for package @cognite/sdk-beta
+
+
+
+
+
+## [6.0.1](https://github.com/cognitedata/cognite-sdk-js/compare/@cognite/sdk-beta@6.0.0...@cognite/sdk-beta@6.0.1) (2025-09-11)
+
+**Note:** Version bump only for package @cognite/sdk-beta
+
+
+
+
+
 # [6.0.0](https://github.com/cognitedata/cognite-sdk-js/compare/@cognite/sdk-beta@5.12.3...@cognite/sdk-beta@6.0.0) (2025-02-03)